+++ src/lib/http/HttpClient.h
@@ -1,3 +1,3 @@
 
 // synchronous request
-int http_request(std::string domain, std::string path, int port, std::string& ret);
+int http_request(std::string domain, std::string path, unsigned short port, std::string& ret);
